% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{bas_report}
\alias{bas_report}
\title{Basal report}
\usage{
bas_report(data, file)
}
\arguments{
\item{data}{}

\item{file}{}
}
\value{
data
}
\description{
Can be used with each specific measurement of basal, cmic and mgrowth.
PDF format is recommended for saving multiple pages. Otherwise other \code{ggsave} formats should work,
but will only save the first page.
}
\examples{
o2meas \%>\% 
  bas_report("bas.pdf") \%>\% 
  cmic_report("cmic.pdf") \%>\% 
  mgrow_report("mgrow.pdf")
}
