% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_writing.R
\name{calc_hour_diffs}
\alias{calc_hour_diffs}
\title{Calculate hourly differences}
\usage{
calc_hour_diffs(x)
}
\arguments{
\item{x}{numeric}
}
\value{
values
}
\description{
Uses the difference between every fourth value, similar to calculations
using the Excel template.
}
