% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_writing.R
\name{check_empty}
\alias{check_empty}
\title{Check if the data is empty}
\usage{
check_empty(ws)
}
\arguments{
\item{ws}{data.frame containing weighing data}
}
\value{
logical
}
\description{
Check if the data is empty
}
