% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_writing.R
\name{check_w_format}
\alias{check_w_format}
\title{Check if the format fits expectations}
\usage{
check_w_format(ws)
}
\arguments{
\item{ws}{data.frame}
}
\value{
logical
}
\description{
Currently fairly conservative, would need to be adapted to fit more funky files
}
