% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating.R
\name{combine_name}
\alias{combine_name}
\title{Combine names}
\usage{
combine_name(n1, n2, n3, symbol = "_")
}
\arguments{
\item{n1}{First name}

\item{n2}{Second name}

\item{n3}{Third name}

\item{symbol}{Used to paste names together}
}
\value{
character
}
\description{
Combine names
}
