% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_writing.R
\name{create_weighing_sheet}
\alias{create_weighing_sheet}
\title{Create weighing sheet from dataframe}
\usage{
create_weighing_sheet(data, input_name, output_dir)
}
\arguments{
\item{data}{data.frame}

\item{input_name}{character file name}

\item{output_dir}{path}
}
\value{
TRUE if file creation worked, NULL if the data is empty
}
\description{
To be used to reformat weighing sheets.
}
