% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{get_set_times}
\alias{get_set_times}
\title{Get set times}
\usage{
get_set_times(id, tib)
}
\arguments{
\item{id}{character}

\item{tib}{tibble with updated times to use}
}
\value{
times
}
\description{
Get set times
}
