% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating.R
\name{o2_bas}
\alias{o2_bas}
\title{Title}
\usage{
o2_bas(data, plot = FALSE, only_sets = FALSE)
}
\arguments{
\item{data}{}

\item{only_sets}{logical if TRUE will only recalculate basal measurements for entries that have bas_set value}
}
\value{

}
\description{
Title
}
