% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating.R
\name{o2_cmic}
\alias{o2_cmic}
\title{Calculate Cmic}
\usage{
o2_cmic(data, plot = FALSE, only_sets = FALSE)
}
\arguments{
\item{data}{dataset}

\item{plot}{logical}

\item{only_sets}{logical if TRUE will only recalculate basal measurements for entries that have bas_set value}
}
\value{
tibble
}
\description{
Calculate Cmic
}
