% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating.R
\name{o2_corr_fct}
\alias{o2_corr_fct}
\title{Correction factors}
\usage{
o2_corr_fct(dev, seq, meas)
}
\arguments{
\item{dev}{Device}

\item{seq}{Sequence}

\item{meas}{Measurements type. "bas" or "mic"}
}
\value{
a number
}
\description{
Returns the  correction factor from list
}
