% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating.R
\name{o2_mgrowth}
\alias{o2_mgrowth}
\title{Calculate Microbial growth}
\usage{
o2_mgrowth(data, plot = FALSE)
}
\arguments{
\item{data}{dataset}

\item{plot}{logical}

\item{only_sets}{logical if TRUE will only recalculate basal measurements for entries that have bas_set value}
}
\value{
tibble
}
\description{
Calculate Microbial growth
}
