% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating.R
\name{o2_process_all}
\alias{o2_process_all}
\title{Process all}
\usage{
o2_process_all(
  files,
  raw_files = "from_wei",
  combine_names = "_",
  plot = FALSE
)
}
\arguments{
\item{files}{}

\item{raw_files}{}

\item{combine_names}{}

\item{plot}{}
}
\value{

}
\description{
Wrapper function to run all analyses with defaults
}
