% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating.R
\name{o2_process_from_table1}
\alias{o2_process_from_table1}
\title{Process O2 table}
\usage{
o2_process_from_table1(
  w_file,
  raw_files = "from_wei",
  combine_names = "_",
  plot = FALSE
)
}
\arguments{
\item{w_file}{}

\item{plot}{}
}
\value{
tibble
}
\description{
The function looks for the raw data files in the same directory
}
