% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_writing.R
\name{o2_read_machine_file}
\alias{o2_read_machine_file}
\title{Read Machine File}
\usage{
o2_read_machine_file(path)
}
\arguments{
\item{path}{to the file}
}
\value{
tibble
}
\description{
Read Machine File
}
