% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_1bas}
\alias{plot_1bas}
\title{Plot 1 Basal}
\usage{
plot_1bas(data, threshold = 3.2)
}
\arguments{
\item{data}{data}

\item{threshold}{numeric Value to cap y axis.}
}
\value{
ggplot
}
\description{
Plot 1 Basal
}
