% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_writing.R
\name{reformat_weighing_sheet}
\alias{reformat_weighing_sheet}
\title{Reformat weighing sheet}
\usage{
reformat_weighing_sheet(file)
}
\arguments{
\item{file}{path}
}
\value{
invisible(file)
}
\description{
Updated based on current format used in 2022-09
}
