% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_writing.R
\name{reformat_weish}
\alias{reformat_weish}
\title{Reformat weighing sheet}
\usage{
reformat_weish(df, info = NULL, template = "iSBio")
}
\arguments{
\item{df}{data}

\item{info}{...}

\item{template}{file}
}
\description{
Project specific, not needed
NOTE: is template needed?
Changed the weighing sheet format to match the template used
}
