% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating.R
\name{sel_mgrowth_df}
\alias{sel_mgrowth_df}
\title{Prepare table for microbial growth calculation}
\usage{
sel_mgrowth_df(data)
}
\arguments{
\item{data}{dataset}
}
\value{
tibble
}
\description{
Prepare table for microbial growth calculation
}
