% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{set_bas_times}
\alias{set_bas_times}
\title{Set basal times manually}
\usage{
set_bas_times(data, tib)
}
\arguments{
\item{data}{dataset}

\item{tib}{tibble with columns "name_c" and "times".}
}
\description{
Adjust the calculation period for soil basal respiration manually.
The column "times" in tib should be a vector of full hours to select
}
\note{
Currently only works with merged names (column name_c).
Duplicated sample names are not handled.
}
