% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_writing.R
\name{write_xl_format}
\alias{write_xl_format}
\title{Title}
\usage{
write_xl_format(df, path)
}
\arguments{
\item{df}{data}

\item{path}{path}
}
\description{
NOTE needs reimplementation
Write Excel file and format
}
