% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_writing.R
\name{check_wei_format}
\alias{check_wei_format}
\title{Check weighing sheet format}
\usage{
check_wei_format(path)
}
\arguments{
\item{path}{}
}
\value{
character indicating format used
}
\description{
Either all in one ws or split on two
}
